+++ src/sfCommand.ts
@@ -6,7 +6,7 @@
  */
 import * as os from 'os';
 import { CliUx, Command, Config, HelpSection, Interfaces } from '@oclif/core';
-import { envVars, Messages, SfdxProject, Mode } from '@salesforce/core';
+import { envVars, Messages, SfdxProject, Lifecycle, Mode } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import chalk from 'chalk';
 import { Progress, Prompter, Spinner, Ux } from './ux';
@@ -45,6 +45,7 @@ export abstract class SfCommand<T> extends Command {
   private warnings: SfCommand.Warning[] = [];
   private ux: Ux;
   private prompter: Prompter;
+  private lifecycle: Lifecycle;
 
   protected get statics(): typeof SfCommand {
     return this.constructor as typeof SfCommand;
@@ -57,6 +58,7 @@ export abstract class SfCommand<T> extends Command {
     this.progress = new Progress(outputEnabled);
     this.ux = new Ux(outputEnabled);
     this.prompter = new Prompter();
+    this.lifecycle = Lifecycle.getInstance();
   }
 
   /**
@@ -150,6 +152,9 @@ export abstract class SfCommand<T> extends Command {
     if (this.statics.requiresProject) {
       this.project = await this.assignProject();
     }
+    this.lifecycle.onWarning(async (warning: string) => {
+      this.warn(warning);
+    });
     return super._run<R>();
   }